requirements.txt
arrow==1.2.3
asttokens==2.0.8
backcall==0.2.0
binaryornot==0.4.4
catboost==1.1
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cookiecutter==2.1.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
fonttools==4.37.4
graphviz==0.20.1
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.4.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.4
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.0
six==1.16.0
sklearn==0.0
src==0.1.0
stack-data==0.5.1
tenacity==8.1.0
text-unidecode==1.3
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5